// src/api/generated/models/operationalReportResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */
import type { OperationalReportResponsePayload } from './operationalReportResponsePayload';

export interface OperationalReportResponse {
  id?: string;
  caseFileId?: string;
  territorialUnitId?: string;
  operationalUnitId?: string;
  templateId?: string;
  source?: string;
  status?: string;
  incidentDate?: string;
  municipalityCode?: string;
  operationType?: string;
  payload?: OperationalReportResponsePayload;
  submittedBy?: string;
  submittedAt?: string;
  validatedBy?: string;
  validatedAt?: string;
  rejectionReason?: string;
}